+++ ChangeLog.md
@@ -16,6 +16,7 @@ Incoming
     * Added `Pattern::template()->literal()`, which is a shorthand for
       `Pattern::template()->builder()->literal()->build()`.
     * Added `Pattern::template()->mask()`, which is a shorthand for `Pattern::template()->builder()->mask()->build()`.
+    * Casting `PatternInterface` to string results in a delimited pattern
 * Breaking changes
     * `match()->getIterator()` no longer preserves the keys of values (like `all()`)
     * `match()->group()->getIterator()` no longer preserves the keys of values (like `all()`)